Add more logic for webview detection and consolidate into a single utils script
